Refactored clock
