+++ components/index.js
@@ -1,8 +1,17 @@
+import CountDown from "./CountDown";
+import Clock from "./Clock";
 import Effect from "./effect";
 import Link from "./link";
-import Route from "./route";
 import ColorPad from "./ColorPad";
 import PrecentageBar from "./PrecentageBar";
 import PlayerController from "./PlayerController";
 
-export { ColorPad, Effect, Link, Route, PrecentageBar, PlayerController };
+export {
+  CountDown,
+  Clock,
+  ColorPad,
+  Effect,
+  Link,
+  PrecentageBar,
+  PlayerController,
+};
