+++ views/Game/index.tsx
@@ -1,45 +1,9 @@
 import { View, Text,SafeAreaView, StyleSheet } from "react-native";
 import { Screen, Player } from "../../constants";
-import { useEffect, useState, useRef } from "react";
-import * as Animatable from 'react-native-animatable';
+import { useEffect, useState } from "react";
 import {startSound,normalSound,clockSound} from "../../helpers/SoundPlayer";
-import {PlayerController} from "../../components";
+import {PlayerController,CountDown,Clock} from "../../components";
 
-const CountDown = ({active,countDown}) => {
-  const countDownRef =  useRef<Animatable.View & View>(null);
-  const flipEven = {
-    0: {
-      transform: [{rotateZ: "0deg"}],
-    },
-    1: {
-      transform: [{rotateZ: "180deg"}],
-    },
-  };
-  const flipOdd = {
-    0: {
-      transform: [{rotateZ: "180deg"}],
-    },
-    1: {
-      transform: [{rotateZ: "360deg"}],
-    },
-  };
-  useEffect(() => {
-    setTimeout(()=> {
-      if (countDown == 0) {
-        countDownRef.current && countDownRef.current.flipOutY();
-      } else {
-        countDownRef.current.animate(countDown % 2 ? flipEven : flipOdd);
-      }
-    }, 1000);
-  }, [countDown]);
-  return (
-    <View style={{display: !active ? "flex" : "none",position: "absolute",width:"100%",height:"100%",alignItems:"center",justifyContent:"center"}}>
-    <Animatable.View ref={countDownRef} animation="flipInY" iterationCount={1} style={{width: Screen.Width * 0.9,height:Screen.Width * 0.9,backgroundColor:"black",borderRadius: Screen.Width * 0.9,display: "flex",alignItems:"center",justifyContent:"center"}}>
-      <Animatable.Text duration={2000} iterationCount={4} style={{fontSize: 300,color: "white"}}>{countDown}</Animatable.Text>
-    </Animatable.View>
-  </View>
-  )
-}
 
 
 
@@ -74,8 +38,8 @@ export default function Game({setCurrentPath}:any) {
         setGameState({...gameState,mode: ""})
       }
       if (countDown == 0) {
-        setTimeout(() => {
         setGameTime(60);
+        setTimeout(() => {
         setActive(true);
       },1000);
       } else {
@@ -87,7 +51,7 @@ export default function Game({setCurrentPath}:any) {
     if(countDown == 0){
       setTimeout(()=> {
         if (gameTime == 0) {
-          setCurrentPath("score");
+          setCurrentPath("Score");
         } else {
           setGameTime(gameTime - 1);
         }
@@ -127,7 +91,7 @@ export default function Game({setCurrentPath}:any) {
         style={styles.innerWrapper}
       >
       <View style={styles.middleWrapper}>
-          <Text style={styles.middleWrapperText}>{gameTime}</Text>
+          <Clock time={gameTime}/>
       </View>
 
         <View style={styles.playerWrapper}>
